Fix import and export errors
