+++ api/stats_test.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"fmt"
